// src/app/interfaces/admin-interfaces/country.interface.ts
// Generated by https://quicktype.io

export interface ICountryListResponse {
  data:  ICountry[];
  total: number;
}

export interface ICountry {
  status:         boolean;
  userCreate:     null;
  createAt:       null;
  id:             string;
  code:           string;
  name:           string;
  continent:      string;
  region:         string;
  surfaceArea:    string;
  indepYear:      number | null;
  population:     number;
  lifeExpectancy: null | string;
  gNP:            string;
  gNPOld:         null | string;
  localName:      string;
  governmentForm: string;
  headOfState:    null | string;
  capital:        number | null;
  code2:          string;
}