// src/app/interfaces/admin-interfaces/timezone.interface.ts
// Generated by https://quicktype.io

export interface ITimezoneListresponse {
  data:  ITimezone[];
  total: number;
}

export interface ITimezone {
  status:       boolean;
  userCreate:   null;
  createAt:     null;
  id:           string;
  zoneName:     string;
  countryCode:  string;
  abbreviation: string;
  timeStart:    string;
  gmtOffset:    number;
  dst:          string;
}